@@ -159,21 +291,22 @@
     "test": "node ./out/test/runTest.js"
   },
   "devDependencies": {
-    "@types/glob": "^7.1.4",
-    "@types/lodash": "^4.14.178",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "14.x",
-    "@types/uuid": "^8.3.3",
-    "@types/vscode": "^1.63.0",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
+    "@types/glob": "^7.2.0",
+    "@types/lodash": "^4.14.186",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "~18.8.0",
+    "@types/uuid": "^8.3.4",
+    "@types/vscode": "^1.71.0",
+    "@typescript-eslint/eslint-plugin": "^5.39.0",
+    "@typescript-eslint/parser": "^5.39.0",
     "@vscode/test-electron": "^1.6.2",
-    "eslint": "^8.1.0",
-    "glob": "^7.1.7",
+    "axios": "^0.27.2",
+    "eslint": "^8.24.0",
+    "form-data": "^4.0.0",
+    "glob": "^7.2.3",
     "lodash": "^4.17.21",
-    "mocha": "^9.1.3",
-    "typescript": "^4.4.4",
+    "mocha": "^9.2.2",
+    "typescript": "^4.8.4",
     "uuid": "^8.3.2"
-  },
-  "dependencies": {}
+  }
 }
